+++ src/QMCWaveFunctions/BsplineFactory/BsplineReaderBase.cpp
@@ -17,7 +17,7 @@
  *
  * Implement super function
  */
-#include "QMCWaveFunctions/EinsplineSetBuilder.h"
+#include "EinsplineSetBuilder.h"
 #include "BsplineReaderBase.h"
 #include "OhmmsData/AttributeSet.h"
 #include "Message/CommOperators.h"
