+++ src/QMCWaveFunctions/BsplineFactory/EinsplineSetBuilder_createSPOs.cpp
@@ -17,7 +17,7 @@
 //////////////////////////////////////////////////////////////////////////////////////
 
 
-#include "QMCWaveFunctions/EinsplineSetBuilder.h"
+#include "EinsplineSetBuilder.h"
 #include <PlatformSelector.hpp>
 #include "CPU/e2iphi.h"
 #include "CPU/SIMD/vmath.hpp"
@@ -28,10 +28,10 @@
 #include "Particle/DistanceTable.h"
 #include <fftw3.h>
 #include "Utilities/ProgressReportEngine.h"
-#include "QMCWaveFunctions/einspline_helper.hpp"
-#include "QMCWaveFunctions/BsplineFactory/BsplineReaderBase.h"
-#include "QMCWaveFunctions/BsplineFactory/BsplineSet.h"
-#include "QMCWaveFunctions/BsplineFactory/createBsplineReader.h"
+#include "einspline_helper.hpp"
+#include "BsplineReaderBase.h"
+#include "BsplineSet.h"
+#include "createBsplineReader.h"
 
 #include <array>
 #include <string_view>
