+++ src/QMCWaveFunctions/BsplineFactory/createComplexDouble.cpp
@@ -15,18 +15,18 @@
 #include "CPU/e2iphi.h"
 #include "CPU/SIMD/vmath.hpp"
 #include "Utilities/ProgressReportEngine.h"
-#include "QMCWaveFunctions/EinsplineSetBuilder.h"
-#include "QMCWaveFunctions/BsplineFactory/BsplineSet.h"
-#include "QMCWaveFunctions/BsplineFactory/SplineC2R.h"
-#include "QMCWaveFunctions/BsplineFactory/SplineC2C.h"
-#include "QMCWaveFunctions/BsplineFactory/SplineC2ROMPTarget.h"
-#include "QMCWaveFunctions/BsplineFactory/SplineC2COMPTarget.h"
-#include "QMCWaveFunctions/BsplineFactory/HybridRepCplx.h"
+#include "EinsplineSetBuilder.h"
+#include "BsplineSet.h"
+#include "SplineC2R.h"
+#include "SplineC2C.h"
+#include "SplineC2ROMPTarget.h"
+#include "SplineC2COMPTarget.h"
+#include "HybridRepCplx.h"
 #include <fftw3.h>
-#include "QMCWaveFunctions/einspline_helper.hpp"
-#include "QMCWaveFunctions/BsplineFactory/BsplineReaderBase.h"
-#include "QMCWaveFunctions/BsplineFactory/SplineSetReader.h"
-#include "QMCWaveFunctions/BsplineFactory/HybridRepSetReader.h"
+#include "einspline_helper.hpp"
+#include "BsplineReaderBase.h"
+#include "SplineSetReader.h"
+#include "HybridRepSetReader.h"
 
 namespace qmcplusplus
 {
