+++ src/QMCWaveFunctions/BsplineFactory/createRealDouble.cpp
@@ -13,15 +13,15 @@
 #include "QMCWaveFunctions/BsplineFactory/createBsplineReader.h"
 #include <PlatformSelector.hpp>
 #include "Utilities/ProgressReportEngine.h"
-#include "QMCWaveFunctions/EinsplineSetBuilder.h"
-#include "QMCWaveFunctions/BsplineFactory/BsplineSet.h"
-#include "QMCWaveFunctions/BsplineFactory/SplineR2R.h"
-#include "QMCWaveFunctions/BsplineFactory/HybridRepReal.h"
+#include "EinsplineSetBuilder.h"
+#include "BsplineSet.h"
+#include "SplineR2R.h"
+#include "HybridRepReal.h"
 #include <fftw3.h>
-#include "QMCWaveFunctions/einspline_helper.hpp"
-#include "QMCWaveFunctions/BsplineFactory/BsplineReaderBase.h"
-#include "QMCWaveFunctions/BsplineFactory/SplineSetReader.h"
-#include "QMCWaveFunctions/BsplineFactory/HybridRepSetReader.h"
+#include "einspline_helper.hpp"
+#include "BsplineReaderBase.h"
+#include "SplineSetReader.h"
+#include "HybridRepSetReader.h"
 
 namespace qmcplusplus
 {
