+++ src/QMCWaveFunctions/tests/test_einset.cpp
@@ -17,8 +17,8 @@
 #include "Particle/ParticleSet.h"
 #include "Particle/ParticleSetPool.h"
 #include "QMCWaveFunctions/WaveFunctionComponent.h"
-#include "QMCWaveFunctions/EinsplineSetBuilder.h"
-#include "QMCWaveFunctions/EinsplineSpinorSetBuilder.h"
+#include "BsplineFactory/EinsplineSetBuilder.h"
+#include "BsplineFactory/EinsplineSpinorSetBuilder.h"
 #include <ResourceCollection.h>
 
 #include <stdio.h>
